ci: set conda python to lowest supported version
It's safest to work with the lowest supported version of Python when
developing the framework locally, because then you are not invited
to use syntax that is not supported (e.g. import from
typing_extensions if you want to use Protocol in 3.6).

Chosen Python 3.6.2, because that's the lowest version of Python that
Conda offers om the default channel.